[Refactor] Extract AbstractSqlSampleSubqueryExecutor from FilesSampleSubqueryExecutor
