+++ fe/fe-core/src/main/java/com/starrocks/alter/reshard/presplit/FilesSampleSubqueryExecutor.java
@@ -15,24 +15,11 @@
 package com.starrocks.alter.reshard.presplit;
 
 import com.google.common.annotations.VisibleForTesting;
-import com.google.gson.JsonArray;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonParser;
 import com.starrocks.catalog.Column;
-import com.starrocks.catalog.Variant;
 import com.starrocks.common.StarRocksException;
 import com.starrocks.common.util.SqlUtils;
-import com.starrocks.qe.ConnectContext;
-import com.starrocks.qe.SimpleExecutor;
-import com.starrocks.statistic.StatisticUtils;
-import com.starrocks.thrift.TResultBatch;
-import com.starrocks.thrift.TResultSinkType;
 import com.starrocks.warehouse.cngroup.ComputeResource;
 
-import java.nio.ByteBuffer;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -44,61 +31,17 @@ import java.util.stream.Collectors;
  * Subclasses provide a {@link Source} (FILES property map plus the load's
  * total input byte total and {@link ComputeResource}); everything else —
  * sampling-rate math, SQL synthesis, BE invocation, JSON row decode — is
- * shared.
+ * inherited from {@link AbstractSqlSampleSubqueryExecutor}.
  *
- * <p>The sub-query is submitted through {@link SimpleExecutor#executeDQL}
- * with a {@link ConnectContext} pinned to the load's compute resource, so
+ * <p>The sub-query is submitted through {@link com.starrocks.qe.SimpleExecutor#executeDQL}
+ * with a {@link com.starrocks.qe.ConnectContext} pinned to the load's compute resource, so
  * cluster resources, audit logging, and credential redaction are inherited
  * rather than reimplemented. Result rows arrive as JSON
- * {@code {"data":[...]}} envelopes (HTTP_PROTOCAL sink); each row's single
- * column is fed to {@link Variant#of} using the sort-key column's declared
- * type.
+ * {@code {"data":[...]}} envelopes (HTTP_PROTOCAL sink); each row's projected
+ * columns are fed to {@link com.starrocks.catalog.Variant#of} using their
+ * respective declared types.
  */
-abstract class FilesSampleSubqueryExecutor implements SampleSubqueryExecutor {
-
-    private static final String EXECUTOR_NAME = "TabletPreSplitDataTierFilesSubquery";
-
-    private static final SimpleExecutor PRODUCTION_SIMPLE_EXECUTOR =
-            new SimpleExecutor(EXECUTOR_NAME, TResultSinkType.HTTP_PROTOCAL);
-
-    /** Soft target row count the sampling rate is sized to deliver. */
-    static final int TARGET_SAMPLE_ROW_COUNT = 50_000;
-
-    /**
-     * Coarse FE-side estimate of an average input row's serialized byte size,
-     * used only to convert the FE-known input byte total into a sampling rate
-     * targeting {@link #TARGET_SAMPLE_ROW_COUNT} rows. Wrong by a constant
-     * factor is fine — the SQL {@code LIMIT} caps over-sampling and the byte
-     * cap in {@link ReservoirSampler} bounds FE memory regardless.
-     */
-    static final long AVERAGE_ROW_BYTES_ESTIMATE = 256L;
-
-    /**
-     * Hard ceiling on rows returned to FE, in addition to the rate-derived
-     * Bernoulli filter. Two over-shoot paths are guarded: a small input
-     * pushes the rate to ~1.0, and a narrow-column input delivers more rows
-     * than expected at a given rate. The 4× factor is large enough to absorb
-     * both without truncating an unbiased sample at the cap.
-     */
-    static final int SAMPLE_ROW_HARD_LIMIT = TARGET_SAMPLE_ROW_COUNT * 4;
-
-    /**
-     * Submits a built sampling SELECT under the supplied {@link ComputeResource}
-     * and returns the BE-side result batches. Constructor-injected to keep the
-     * executor unit-testable without bringing up the full FE planner/coordinator
-     * stack — production wires this to {@link SimpleExecutor#executeDQL} via a
-     * scoped {@link ConnectContext} carrying the load's compute resource.
-     *
-     * <p>{@code queryTimeoutSeconds} caps the sub-query's wall-clock runtime
-     * ({@code 0} = uncapped); the production runner applies it via
-     * {@code query_timeout} on the sample context so an over-budget sample is
-     * cancelled by the BE. Test stubs return canned batches and ignore it.
-     */
-    @FunctionalInterface
-    interface SampleQueryRunner {
-        List<TResultBatch> run(String sampleSql, ComputeResource computeResource, int queryTimeoutSeconds)
-                throws StarRocksException;
-    }
+abstract class FilesSampleSubqueryExecutor extends AbstractSqlSampleSubqueryExecutor {
 
     /** Subclass-supplied FILES sub-query inputs. */
     protected record Source(
@@ -112,17 +55,13 @@ abstract class FilesSampleSubqueryExecutor implements SampleSubqueryExecutor {
         }
     }
 
-    private final String errorPrefix;
-    private final SampleQueryRunner sampleQueryRunner;
-
     FilesSampleSubqueryExecutor(String errorPrefix) {
-        this(errorPrefix, FilesSampleSubqueryExecutor::runViaSimpleExecutor);
+        super(errorPrefix);
     }
 
     @VisibleForTesting
     FilesSampleSubqueryExecutor(String errorPrefix, SampleQueryRunner sampleQueryRunner) {
-        this.errorPrefix = Objects.requireNonNull(errorPrefix, "errorPrefix");
-        this.sampleQueryRunner = Objects.requireNonNull(sampleQueryRunner, "sampleQueryRunner");
+        super(errorPrefix, sampleQueryRunner);
     }
 
     /**
@@ -134,91 +73,34 @@ abstract class FilesSampleSubqueryExecutor implements SampleSubqueryExecutor {
     protected abstract Source resolveSource(SampleRequest request) throws StarRocksException;
 
     @Override
-    public final SampleExecution execute(SampleRequest request) throws StarRocksException {
+    protected final SampleSpec resolveSampleSpec(SampleRequest request) throws StarRocksException {
         Source source = resolveSource(request);
+        String fromClauseSql = "FILES(" + buildPropertiesClause(source.filesProperties()) + ")";
         List<Column> sortKeyColumns = request.getSortKey();
         List<Column> partitionSourceColumns = request.getPartitionSourceColumns();
-
-        double samplingRate = pickSamplingRate(source.totalFileBytes());
-        int rowLimit = pickRowLimit(request.getSampleByteLimit());
-
-        String sampleSql = buildSampleSql(
-                source.filesProperties(), sortKeyColumns, partitionSourceColumns,
-                samplingRate, rowLimit, request.getSeed());
-        List<TResultBatch> resultBatches = runSampleQuery(
-                sampleSql, source.computeResource(), request.getQueryTimeoutSeconds());
-        Iterator<SampleRow> rowIterator =
-                decodeRows(resultBatches, sortKeyColumns, partitionSourceColumns).iterator();
-        return new SampleExecution(rowIterator, new Estimates(source.totalFileBytes(), 0L));
+        return new SampleSpec(fromClauseSql, /*whereClauseSqlOrNull=*/ null,
+                source.totalFileBytes(), source.computeResource(),
+                identsOf(sortKeyColumns), identsOf(partitionSourceColumns),
+                sortKeyColumns, partitionSourceColumns);
     }
 
-    private static double pickSamplingRate(long totalFileBytes) {
-        if (totalFileBytes <= 0L) {
-            return 1.0;
-        }
-        double rate = (double) TARGET_SAMPLE_ROW_COUNT * AVERAGE_ROW_BYTES_ESTIMATE / totalFileBytes;
-        return Math.min(1.0, rate);
+    private static List<String> identsOf(List<Column> columns) {
+        return columns.stream()
+                .map(column -> SqlUtils.getIdentSql(column.getName()))
+                .collect(Collectors.toList());
     }
 
     /**
-     * Caps the SQL-side {@code LIMIT} at whichever is smaller: the per-feature
-     * hard cap, or the FE-memory cap implied by {@code sampleByteLimit}. The
-     * byte cap matters because {@code SimpleExecutor.executeDQL} materializes
-     * every result batch on the FE heap before this executor decodes — without
-     * this cap, a rate over-estimate against narrow rows can balloon FE memory
-     * before {@link ReservoirSampler}'s soft byte cap engages.
-     */
-    private static int pickRowLimit(long sampleByteLimit) {
-        long rowLimitFromBytes = sampleByteLimit / AVERAGE_ROW_BYTES_ESTIMATE;
-        return (int) Math.min(SAMPLE_ROW_HARD_LIMIT, Math.max(1L, rowLimitFromBytes));
-    }
-
-    /**
-     * Builds the sampling SELECT against the supplied FILES properties. Projects
-     * every sort-key column followed by every partition-source column (so the
-     * projection list is {@code sortKey, partitionSource} in order) — multi-
-     * column sort keys produce tuple-valued samples that {@link BoundaryPlanner}
-     * can lex-sort and quantile-cut, and the trailing partition-source columns
-     * feed a downstream multi-partition grouper. When
-     * {@code partitionSourceColumns} is empty the projection collapses to the
-     * sort key alone, identical to the pre-extension shape.
-     *
-     * <p>String literal escaping covers both {@code "} and {@code \} so a
+     * Builds the {@code FILES(...)} properties clause from the supplied map.
+     * String literal escaping covers both {@code "} and {@code \} so a
      * crafted property value cannot break out of the double-quoted form and
-     * inject SQL into the internal-context sub-query. The {@code ORDER BY
-     * rand(seed XOR 0x5...)} before {@code LIMIT} re-shuffles the
-     * {@code WHERE}-survivors so an over-rate truncation does not bias
-     * toward earlier files in scan order. When {@code partitionSourceColumns}
-     * is empty the projection collapses to the sort key alone.
+     * inject SQL into the internal-context sub-query.
      */
-    @VisibleForTesting
-    static String buildSampleSql(
-            Map<String, String> filesProperties,
-            List<Column> sortKeyColumns,
-            List<Column> partitionSourceColumns,
-            double samplingRate, int rowLimit, long seed) {
-        List<String> projectedIdentifiers = new ArrayList<>(
-                sortKeyColumns.size() + partitionSourceColumns.size());
-        for (Column column : sortKeyColumns) {
-            projectedIdentifiers.add(SqlUtils.getIdentSql(column.getName()));
-        }
-        for (Column column : partitionSourceColumns) {
-            projectedIdentifiers.add(SqlUtils.getIdentSql(column.getName()));
-        }
-        String projection = String.join(", ", projectedIdentifiers);
-        String propertiesClause = filesProperties.entrySet().stream()
+    static String buildPropertiesClause(Map<String, String> filesProperties) {
+        return filesProperties.entrySet().stream()
                 .map(property -> '"' + escapeDoubleQuoted(property.getKey()) + "\" = \""
                         + escapeDoubleQuoted(property.getValue()) + '"')
                 .collect(Collectors.joining(", "));
-        long orderShuffleSeed = seed ^ 0x5A5A5A5A5A5A5A5AL;
-        return String.format(
-                "SELECT %s FROM FILES(%s) WHERE rand(%d) < %s ORDER BY rand(%d) LIMIT %d",
-                projection,
-                propertiesClause,
-                seed,
-                Double.toString(samplingRate),
-                orderShuffleSeed,
-                rowLimit);
     }
 
     /**
@@ -234,183 +116,4 @@ abstract class FilesSampleSubqueryExecutor implements SampleSubqueryExecutor {
         }
         return value.replace("\\", "\\\\").replace("\"", "\\\"");
     }
-
-    private List<TResultBatch> runSampleQuery(
-            String sampleSql, ComputeResource computeResource, int queryTimeoutSeconds) throws StarRocksException {
-        try {
-            return sampleQueryRunner.run(sampleSql, computeResource, queryTimeoutSeconds);
-        } catch (RuntimeException runtimeFailure) {
-            throw new StarRocksException(
-                    errorPrefix + "sample sub-query failed: " + runtimeFailure.getMessage(), runtimeFailure);
-        }
-    }
-
-    /**
-     * Production sub-query runner: builds a {@link ConnectContext} pinned to
-     * the load's {@link ComputeResource} so the sample executes on the same
-     * warehouse the user load will use, not the statistics-default one
-     * {@link SimpleExecutor#executeDQL(String)} would build implicitly. Thread-
-     * local context handling follows the same prior-context save/restore
-     * pattern {@code SimpleExecutor.executeDQL(String)} uses internally.
-     */
-    private static List<TResultBatch> runViaSimpleExecutor(
-            String sampleSql, ComputeResource computeResource, int queryTimeoutSeconds) {
-        ConnectContext priorContext = ConnectContext.get();
-        ConnectContext sampleContext = configureSampleContext(
-                StatisticUtils.buildConnectContext(), computeResource, queryTimeoutSeconds);
-        sampleContext.setThreadLocalInfo();
-        try {
-            return PRODUCTION_SIMPLE_EXECUTOR.executeDQL(sampleSql, sampleContext);
-        } finally {
-            ConnectContext.remove();
-            if (priorContext != null) {
-                priorContext.setThreadLocalInfo();
-            }
-        }
-    }
-
-    /**
-     * Aligns a freshly built {@link ConnectContext} with the load's compute
-     * resource. The warehouse id MUST be set before
-     * {@code setCurrentComputeResource}: {@code ConnectContext.getCurrentComputeResource}
-     * discards and re-acquires the resource if its {@code getWarehouseId()}
-     * disagrees with the context's {@code currentWarehouseId}, so omitting
-     * the warehouse-id alignment silently routes the sub-query to the
-     * statistics-default warehouse instead of the load's.
-     *
-     * <p>{@code queryTimeoutSeconds > 0} sets {@code query_timeout} on the
-     * sample session so the BE cancels an over-budget sample (the data-tier
-     * pipeline derives this from the remaining pre-submit budget); {@code 0}
-     * leaves the built context's default timeout untouched. This is the seam
-     * that makes the pre-submit deadline hard — {@code executeDQL} reads the
-     * cap from {@code SessionVariable.toThrift()}, not from any SQL hint.
-     */
-    @VisibleForTesting
-    static ConnectContext configureSampleContext(
-            ConnectContext context, ComputeResource computeResource, int queryTimeoutSeconds) {
-        if (queryTimeoutSeconds > 0) {
-            context.getSessionVariable().setQueryTimeoutS(queryTimeoutSeconds);
-        }
-        context.setCurrentWarehouseId(computeResource.getWarehouseId());
-        context.setCurrentComputeResource(computeResource);
-        context.setNeedQueued(false);
-        context.setStartTime();
-        return context;
-    }
-
-    private List<SampleRow> decodeRows(
-            List<TResultBatch> resultBatches,
-            List<Column> sortKeyColumns,
-            List<Column> partitionSourceColumns) throws StarRocksException {
-        List<SampleRow> rows = new ArrayList<>();
-        if (resultBatches == null) {
-            return rows;
-        }
-        for (TResultBatch resultBatch : resultBatches) {
-            List<ByteBuffer> batchRows = resultBatch.getRows();
-            if (batchRows == null) {
-                continue;
-            }
-            for (ByteBuffer rowBuffer : batchRows) {
-                rows.add(decodeRow(rowBuffer, sortKeyColumns, partitionSourceColumns));
-            }
-        }
-        return rows;
-    }
-
-    /**
-     * Decode one HTTP_PROTOCAL JSON row ({@code {"data":[<val0>, ...]}}) into a
-     * {@link SampleRow}. The JSON array carries
-     * {@code sortKeyColumns.size() + partitionSourceColumns.size()} cells in
-     * projection order: the first slice fills the row's sort-key tuple, the
-     * trailing slice fills its partition-source tuple. When
-     * {@code partitionSourceColumns} is empty the trailing slice is empty and
-     * the row collapses to the pre-extension single-tuple shape.
-     *
-     * <p>Nullable columns accept JSON nulls and decode to
-     * {@link com.starrocks.catalog.NullVariant}; non-nullable columns surface a
-     * null cell as an executor failure rather than silently dropping the row.
-     * Any shape deviation or type-coerce failure becomes a
-     * {@link StarRocksException} so the coordinator records SAMPLE_FAILED
-     * instead of letting an unchecked exception unwind the load thread.
-     */
-    private SampleRow decodeRow(
-            ByteBuffer rowBuffer,
-            List<Column> sortKeyColumns,
-            List<Column> partitionSourceColumns) throws StarRocksException {
-        int expectedArity = sortKeyColumns.size() + partitionSourceColumns.size();
-        JsonArray dataArray = extractDataArray(rowBuffer, expectedArity);
-        List<Variant> sortKeyValues = new ArrayList<>(sortKeyColumns.size());
-        for (int columnIndex = 0; columnIndex < sortKeyColumns.size(); columnIndex++) {
-            sortKeyValues.add(decodeCell(
-                    dataArray.get(columnIndex), sortKeyColumns.get(columnIndex), COLUMN_ROLE_SORT_KEY));
-        }
-        List<Variant> partitionSourceValues = new ArrayList<>(partitionSourceColumns.size());
-        for (int columnIndex = 0; columnIndex < partitionSourceColumns.size(); columnIndex++) {
-            partitionSourceValues.add(decodeCell(
-                    dataArray.get(sortKeyColumns.size() + columnIndex),
-                    partitionSourceColumns.get(columnIndex),
-                    COLUMN_ROLE_PARTITION_SOURCE));
-        }
-        return new SampleRow(sortKeyValues, partitionSourceValues);
-    }
-
-    /**
-     * Parse {@code rowBuffer} as JSON and return its {@code data} array,
-     * validating shape (non-empty UTF-8, root is JSON object, has a
-     * {@code data} array, arity matches the expected projection count).
-     */
-    private JsonArray extractDataArray(ByteBuffer rowBuffer, int expectedArity) throws StarRocksException {
-        String jsonRow = StandardCharsets.UTF_8.decode(rowBuffer.duplicate()).toString();
-        JsonElement root;
-        try {
-            root = JsonParser.parseString(jsonRow);
-        } catch (RuntimeException parseFailure) {
-            throw new StarRocksException(
-                    errorPrefix + "row decode failed: " + parseFailure.getMessage(), parseFailure);
-        }
-        if (!root.isJsonObject()) {
-            throw new StarRocksException(errorPrefix + "row root is not a JSON object: " + jsonRow);
-        }
-        JsonElement dataElement = root.getAsJsonObject().get("data");
-        if (dataElement == null || !dataElement.isJsonArray()) {
-            throw new StarRocksException(errorPrefix + "row is missing a JSON array `data` field: " + jsonRow);
-        }
-        JsonArray dataArray = dataElement.getAsJsonArray();
-        if (dataArray.size() != expectedArity) {
-            throw new StarRocksException(errorPrefix + "expected " + expectedArity
-                    + " projected column(s) per row but row carried " + dataArray.size());
-        }
-        return dataArray;
-    }
-
-    /**
-     * Coerce one JSON cell into a {@link Variant} of {@code column}'s declared
-     * type. Null cells produce a typed {@link Variant#nullVariant} for nullable
-     * columns (BoundaryPlanner's compareTo orders {@code NullVariant} lower
-     * than any non-null value); non-nullable columns reject null as a schema
-     * invariant violation. {@code columnRole} names the projection slice the
-     * cell belongs to (sort-key or partition-source) so a non-null violation
-     * reports the schema an operator should actually inspect.
-     */
-    private Variant decodeCell(JsonElement valueElement, Column column, String columnRole) throws StarRocksException {
-        if (valueElement.isJsonNull()) {
-            if (column.isAllowNull()) {
-                return Variant.nullVariant(column.getType());
-            }
-            throw new StarRocksException(errorPrefix + "sample returned a null value for non-nullable "
-                    + columnRole + " column " + column.getName());
-        }
-        try {
-            return Variant.of(column.getType(), valueElement.getAsString());
-        } catch (RuntimeException variantFailure) {
-            throw new StarRocksException(errorPrefix + "failed to coerce sample value for " + columnRole
-                    + " column " + column.getName() + " to " + column.getType().toSql() + ": "
-                    + variantFailure.getMessage(),
-                    variantFailure);
-        }
-    }
-
-    private static final String COLUMN_ROLE_SORT_KEY = "sort-key";
-    private static final String COLUMN_ROLE_PARTITION_SOURCE = "partition-source";
 }
